Fix issues found verifying against a live Supabase project
- board_state(): jsonb ->> only accepts int/text, not the bigint
  WITH ORDINALITY produces (0002 patches the already-deployed function,
  0001 carries the fix for anyone provisioning from scratch)
- Route admin/company creation through a new create-company Edge
  Function using the Auth Admin API, same as player creation - the
  public signUp() endpoint applies stricter anti-bot email validation
  and rate limits that don't make sense for these admin-provisioned
  synthetic accounts, and were rejecting every signup attempt
- Switch synthetic account emails from a .local domain to a .com-style
  one - Supabase Auth rejects .local as an invalid TLD
- Refetch launchedCompanies when opening "Play a Campaign" instead of
  only at mount, so a campaign launched earlier in the same session
  actually shows up

Verified end-to-end against the live battleship-incentives Supabase
project: create company -> board setup -> add player -> launch ->
player login -> forced password change -> winning play -> admin Live
Monitor, confirming ship data never reaches the player's browser
(only board_state()/play_move() responses, never raw ships rows).

+++ supabase/migrations/0001_init.sql
@@ -267,7 +267,7 @@ begin
       sh.sunk,
       (t.elem ->> 'r')::int as r,
       (t.elem ->> 'c')::int as c,
-      (sh.hits ->> (t.idx - 1))::boolean as is_hit
+      (sh.hits ->> (t.idx - 1)::int)::boolean as is_hit
     from ships sh
     cross join lateral jsonb_array_elements(sh.cells) with ordinality as t(elem, idx)
     where sh.company_id = p_company_id
